requirements.txt
python_stopwatch==1.1.11
Requests==2.32.5
selenium==4.35.0
selenium_wire==5.1.0
sqlparse==0.5.3
